requirements.txt
beautifulsoup4>=4.12.0
pyyaml>=6.0
pytest>=7.4.0
lxml>=4.9.0
